Refactor to HYBRID-only: remove ONEDNN files, move build_sdpa() into hybrid
- Remove fattn-onednn.cpp/hpp (reserved for PR #25222)
- Move sdpa_partition + build_sdpa() into fattn-hybrid.cpp/hpp
- Strip FA-DISP diagnostics from fattn.cpp
- Comment out ONEDNN enum/dispatch (uncomment when #25222 merges)
- Update gate comments for HYBRID-only dispatch order
- HYBRID now ships build_sdpa() independently — no dependency on ONEDNN files

Build: cmake --preset x64-windows-sycl-release -DGGML_SYCL_DNN=ON -DGGML_SYCL_F16=ON
